Designing a Store Database
This is the first code that I wrote in SQL. The lesson of this manner was to create a table that represents the average price of purchasing an new car. To add on, the CREATE TABLE  function is to create a table, but you need a value which in this case is cars. Then I used the id integer primary key, name text, prices integer, to create columns to be visual while creating the code from the sequence.  After creating the table, then I had to INSERT INTO (statement) which in this example would be cars to plug in VALUES which input the model of the vehicle, prices and what column there in to look up that data. In this example, I wanted to know the average price of all the vehicles to see  how much money I would need to buy a brand new car. In greater detail, I used the SELECT AVG (prices) FROM cars; TO calculate the average price of a new vehicle.